sửa giao diện trang quanly_vienchuc_khoa

+++ app/Http/Controllers/VienChucController.php
@@ -410,7 +410,28 @@ class VienChucController extends Controller
       return Redirect::to('/home');
     }
   }
-  public function admin_delete_vienchuc($ma_vc){
+  // public function admin_delete_vienchuc($ma_vc){
+  //   $this->check_login();
+  //   $ma_vc_login = session()->get('ma_vc');
+  //   $phanquyen_admin = PhanQuyen::where('ma_vc', $ma_vc_login)
+  //     ->where('ma_q', '=', '5')
+  //     ->first();
+  //   $phanquyen_qlk = PhanQuyen::where('ma_vc', $ma_vc_login)
+  //     ->where('ma_q', '=', '9')
+  //     ->first();
+  //   if($phanquyen_admin || $phanquyen_qlk){
+  //     VienChuc::find($ma_vc)->delete();
+  //     if($phanquyen_qlk){
+  //       return Redirect::to('thongtin_vienchuc_khoa');
+  //     }else{
+  //       return Redirect::to('quanly_vienchuc_khoa');
+  //     }
+  //   }else{
+  //     return Redirect::to('/home');
+  //   }
+  // }
+  public function admin_delete_vienchuc(Request $request){
+    $this->check_login();
     $this->check_login();
     $ma_vc_login = session()->get('ma_vc');
     $phanquyen_admin = PhanQuyen::where('ma_vc', $ma_vc_login)
@@ -420,14 +441,12 @@ class VienChucController extends Controller
       ->where('ma_q', '=', '9')
       ->first();
     if($phanquyen_admin || $phanquyen_qlk){
-      VienChuc::find($ma_vc)->delete();
-      if($phanquyen_qlk){
-        return Redirect::to('thongtin_vienchuc_khoa');
-      }else{
-        return Redirect::to('quanly_vienchuc_khoa');
+      if($request->ajax()){
+        $ma_vc =$request->ma_vc;
+        if($ma_vc != null){
+          VienChuc::find($ma_vc)->delete();
+        }
       }
-    }else{
-      return Redirect::to('/home');
     }
   }
   public function search_vienchuc_khoa($ma_k){
